♻️ refactor[#22] : user-auth 병합. - 컨트롤러, 서비스 클래스는 따로 분리. 데이터베이스는 공유
