--- com.sparta.logistics.auth-service/src/main/java/com/sparta/logistics/auth_service/application/dto/request/AuthSignupRequestDto.java
@@ -1,12 +1,0 @@
-package com.sparta.logistics.auth_service.application.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class AuthSignupRequestDto {
-
-    private String username;
-    private String password;
-    private String slackId;
-
-}
