--- com.sparta.logistics.auth-service/src/main/java/com/sparta/logistics/auth_service/domain/repository/AuthRepository.java
@@ -1,12 +1,0 @@
-package com.sparta.logistics.auth_service.domain.repository;
-
-import com.sparta.logistics.auth_service.domain.entity.Auth;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface AuthRepository {
-
-    Optional<Auth> findById(UUID id);
-    Auth save(Auth auth);
-
-}
