--- com.sparta.logistics.auth-service/src/main/java/com/sparta/logistics/auth_service/infrastructure/dto/request/UserSignupRequestDto.java
@@ -1,15 +1,0 @@
-package com.sparta.logistics.auth_service.infrastructure.dto.request;
-
-import java.util.UUID;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserSignupRequestDto {
-
-    private UUID authId;
-    private String username;
-    private String slackId;
-
-}
