--- com.sparta.logistics.auth-service/src/main/java/com/sparta/logistics/auth_service/infrastructure/dto/response/UserSignupResponseDto.java
@@ -1,12 +1,0 @@
-package com.sparta.logistics.auth_service.infrastructure.dto.response;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserSignupResponseDto {
-
-    private String username;
-    private String slackId;
-
-}
