--- com.sparta.logistics.auth-service/src/main/java/com/sparta/logistics/auth_service/infrastructure/jpa/AuthJpaRepository.java
@@ -1,11 +1,0 @@
-package com.sparta.logistics.auth_service.infrastructure.jpa;
-
-import com.sparta.logistics.auth_service.domain.entity.Auth;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuthJpaRepository extends JpaRepository<Auth, UUID> {
-
-}
