+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/response/UserPasswordResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.logistics.user_service.presentation.dto.response;
+package com.sparta.logistics.user_service.application.dto.response;
 
 import lombok.Data;
