+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/dto/response/UserUpdateResponseDto.java
@@ -1,0 +1,8 @@
+package com.sparta.logistics.user_service.application.dto.response;
+
+import lombok.Data;
+
+@Data
+public class UserUpdateResponseDto {
+
+}
