+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/application/service/UserService.java
@@ -1,0 +1,13 @@
+package com.sparta.logistics.user_service.application.service;
+
+import com.sparta.logistics.user_service.domain.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+
+}
