+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/infrastructure/repository/UserRepositoryImpl.java
@@ -13,7 +13,6 @@ public class UserRepositoryImpl implements UserRepository {
 
     private final UserJpaRepository userJpaRepository;
 
-
     @Override
     public Optional<User> findById(Long id) {
         return userJpaRepository.findById(id);
