+++ com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/jwt/JwtUtil.java
@@ -1,11 +1,12 @@
-package com.sparta.logistics.auth_service.jwt;
+package com.sparta.logistics.user_service.jwt;
+
+
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.SignatureException;
 import io.jsonwebtoken.UnsupportedJwtException;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
@@ -73,7 +74,7 @@ public class JwtUtil {
         try {
             Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
             return true;
-        } catch (SecurityException | MalformedJwtException | SignatureException e) {
+        } catch (SecurityException | MalformedJwtException e) {
             log.error("Invalid JWT signature");
         } catch (ExpiredJwtException e) {
             log.error("Expired JWT token");
