--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/presentation/dto/request/UserDeleteRequestDto.java
@@ -1,8 +1,0 @@
-package com.sparta.logistics.user_service.presentation.dto.request;
-
-import lombok.Data;
-
-@Data
-public class UserDeleteRequestDto {
-
-}
