--- com.sparta.logistics.user-service/src/main/java/com/sparta/logistics/user_service/presentation/dto/response/UserUpdateResponseDto.java
@@ -1,8 +1,0 @@
-package com.sparta.logistics.user_service.presentation.dto.response;
-
-import lombok.Data;
-
-@Data
-public class UserUpdateResponseDto {
-
-}
